fix(export): overflow when correcting number of frames

+++ Holovibes/includes/io_files/output_avi_file.hh
@@ -66,7 +66,7 @@ class OutputAviFile : public OutputFrameFile, public AviFile
      *  \throw    FileException if an error occurred while correcting the
      * sections
      */
-    void correct_number_of_frames(unsigned int nb_frames_written) override;
+    void correct_number_of_frames(size_t nb_frames_written) override;
 
   private:
     // Give access to private members to the factory
